관리자 페이지, ID 확인, 관리자라도 본인 것만 삭제

+++ src/components/AdminList.js
@@ -6,8 +6,8 @@
 import React, { useState, useEffect } from "react";
 import { useSession, signOut } from "next-auth/react";
 
-import TodoItem from "@/components/TodoItem";
-import styles from "@/styles/TodoList.module.css";
+import AdminItem from "./AdminItem";
+import styles from "@/styles/AdminList.module.css";
 
 // firebase 관련 모듈을 불러옵니다.
 import { db } from "@/firebase";
@@ -16,7 +16,6 @@ import {
   query,
   doc,
   getDocs,
-  addDoc,
   updateDoc,
   deleteDoc,
   orderBy,
@@ -27,11 +26,9 @@ import {
 const todoCollection = collection(db, "todos");
 
 // TodoList 컴포넌트를 정의합니다.
-const TodoList = () => {
+const AdminList = () => {
   // 상태를 관리하는 useState 훅을 사용하여 할 일 목록과 입력값을 초기화합니다.
   const [todos, setTodos] = useState([]);
-  const [input, setInput] = useState("");
-
   const { data } = useSession();
 
   const getTodos = async () => {
@@ -43,7 +40,6 @@ const TodoList = () => {
 
     const q = query(
       todoCollection,
-      where("userId", "==", data?.user?.id),
       orderBy("timestamp", "asc")
     );
 
@@ -65,51 +61,12 @@ const TodoList = () => {
     getTodos();
   }, [data]);
 
-  // addTodo 함수는 입력값을 이용하여 새로운 할 일을 목록에 추가하는 함수입니다.
-  const addTodo = async () => {
-    // 입력값이 비어있는 경우 함수를 종료합니다.
-    if (input.trim() === "") return;
-    // 기존 할 일 목록에 새로운 할 일을 추가하고, 입력값을 초기화합니다.
-    // {
-    //   id: 할일의 고유 id,
-    //   text: 할일의 내용,
-    //   completed: 완료 여부,
-    // }
-    // ...todos => {id: 1, text: "할일1", completed: false}, {id: 2, text: "할일2", completed: false}}, ..
-
-    // Firestore 에 추가한 할 일을 저장합니다.
-    const docRef = await addDoc(todoCollection, {
-      userId: data?.user?.id,
-      userName: data?.user?.name,
-      text: input,
-      completed: false,
-      timestamp: Date.now(),
-    });
-
-    // id 값을 Firestore 에 저장한 값으로 지정합니다.
-    setTodos([...todos, {
-      id: docRef.id, 
-      userId: data?.user?.id,
-      userName: data?.user?.name,
-      text: input,
-      completed: false,
-      timestamp: Date.now()
-    }]);
-    setInput("");
-    const container = document.querySelector(`.${styles.container}`);
-    container.classList.add("animate__animated", "animate__bounce");
-    setTimeout(() => {
-      container.classList.remove("animate__animated", "animate__bounce");
-    }, 1000);
-  };
-
-
   
   // toggleTodo 함수는 체크박스를 눌러 할 일의 완료 상태를 변경하는 함수입니다.
   const toggleTodo = (id) => {
     // 할 일 목록에서 해당 id를 가진 할 일의 완료 상태를 반전시킵니다.
     const newTodos = todos.map((todo) => {
-      if (todo.id === id && todo.userId === data?.user?.id) {
+      if (todo.id === id) {
         // Firestore 에서 해당 id를 가진 할 일을 찾아 완료 상태를 업데이트합니다.
         const todoDoc = doc(todoCollection, id);
         updateDoc(todoDoc, { completed: !todo.completed });
@@ -124,8 +81,13 @@ const TodoList = () => {
   };
 
   // deleteTodo 함수는 할 일을 목록에서 삭제하는 함수입니다.
-  const deleteTodo = (id) => {
+  const deleteTodo = (id, userId) => {
     // Firestore 에서 해당 id를 가진 할 일을 삭제합니다.
+    if (userId !== data?.user.id) {
+      alert("권한이 없습니다.");
+      return;
+    }
+
     const todoDoc = doc(todoCollection, id);
     deleteDoc(todoDoc);
 
@@ -140,12 +102,14 @@ const TodoList = () => {
 
   // 컴포넌트를 렌더링합니다.
   return (
+    // 관리자 페이지로 변경
     <div className={styles.container}>
       <div className="flex items-center justify-between mb-4">
         <h1 className="text-xl mb-4 font-bold" style={{ 
-        textDecoration: "underline dotted", color: "gray" }}>
-          <span className="font-bold">{data?.user?.name}</span>의 Todo List
+        textDecoration: "underline dotted", color: "blue" }}>
+          Admin Page
         </h1>
+
         {/* 로그아웃 버튼 추가 */}
         {data?.user?.name && (
           <button
@@ -156,58 +120,17 @@ const TodoList = () => {
           </button>
         )}
       </div>
-      
-      {/* 할 일을 입력받는 텍스트 필드입니다. */}
-      <input
-        type="text"
-        // className={styles.itemInput}
-        // -- itemInput CSS code --
-        // input[type="text"].itemInput {
-        //   width: 100%;
-        //   padding: 5px;
-        //   margin-bottom: 10px;
-        // }
-        className="shadow-lg w-full p-1 mb-4 border border-gray-300 rounded"
-        value={input}
-        onChange={(e) => setInput(e.target.value)}
-        onKeyDown={(e) => {
-          if (e.key === "Enter") {
-            addTodo();
-          }
-        }}
-      />
-      {/* 할 일을 추가하는 버튼입니다. */}
-      <div className="grid">
-        <button
-          // className={styles.addButton}
-          // -- addButton CSS code --
-          // button.addButton {
-          //   padding: 5px;
-          //   background-color: #0070f3;
-          //   color: white;
-          //   border: 1px solid #0070f3;
-          //   border-radius: 5px;
-          //   cursor: pointer;
-          // }
-          //
-          // button.addButton:hover {
-          //   background-color: #fff;
-          //   color: #0070f3;
-          // }
-          className="shadow-lg w-40 justify-self-end p-1 mb-4 font-bold bg-blue-500 text-white border border-blue-500 rounded hover:bg-white hover:text-blue-500"
-          onClick={addTodo}
-        >
-          Add Todo
-        </button>
-      </div>
+        
+
       {/* 할 일 목록을 렌더링합니다. */}
       <ul>
         {todos.map((todo) => (
-          <TodoItem
+          <AdminItem
             key={todo.id}
             todo={todo}
             onToggle={() => toggleTodo(todo.id)}
             onDelete={() => deleteTodo(todo.id, todo.userId)}
+            userName={todo.userName}
           />
         ))}
       </ul>
@@ -215,4 +138,4 @@ const TodoList = () => {
   );
 };
 
-export default TodoList;
+export default AdminList;
